Isolate functions only used by tests
