+++ javascript/src/tests/testutils.ts
@@ -1,0 +1,40 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {parseHTML} from 'linkedom';
+
+/**
+ * Sets an innerHTML on a given Element or ShadowRoot.
+ * @param element An Element or ShadowRoot.
+ * @param html An HTML string to set.
+ */
+export const setInnerHtml = (element: Element | ShadowRoot, html: string) => {
+  element.innerHTML = html;
+};
+
+/**
+ * Creates an HTML document.
+ * @returns Document
+ */
+export const createDocument = () => {
+  const {document} = parseHTML('<!doctype html><html></html>');
+  return document;
+};
+
+/**
+ * Whether the running environment is a Web browser.
+ */
+export const isBrowser = false;
